requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
setuptools==76.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
